requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
discord.py==2.5.2
dotenv==0.9.9
frozenlist==1.6.0
greenlet==3.2.2
idna==3.10
multidict==6.4.4
numpy==2.2.6
pandas==2.2.3
pip==21.2.4
playwright==1.52.0
propcache==0.3.1
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
setuptools==58.1.0
six==1.17.0
typing_extensions==4.13.2
tzdata==2025.2
yarl==1.20.0